Install Gulp, including the Babel ES2015 preset
Gulp will automatically pre-process its files using Babel and the ES2015
preset. That is, the gulpfile can be written using ES2015, as can spec
files to be fed to Mocha, provided the test runner will be invoked from
a gulp task.

Add, clarify, and re-order the instructions for setting up an
appropriate local development environment for code contributions.