Create a minimal Express app
Install npm packages:
* Express: a web framework for Node.js
* Chai-as-Promised: extends Chai with assertions about promises

Define a minimal Express app that simply responds to 'GET /' with status
'200 OK'. There is also a script at 'bin/www' which launches the server.

Alter 'dummy.spec.js' to set all of its tests to 'pending' to at least
keep it out of the way until it's finally deleted.