Provide AuthToken service to manage an authentication JWT
AuthToken can save, read, erase, and decode a JSON Web Token (JWT) in
the browser's window.localStorage. The key name can be configured via a
setter on the provider, however, a sensible default is pre-configured as
a constant on the module.

decode() expects the JWT to conform to IETF RFC 7519 for use as a bearer
token as described in RFC 6750, (and further in RFCs 7235 and 7617). It
returns the un-encoded payload as a JSON object.

Also, delete the superfluous bogus spec used to test the initial karma
configuration.