Respond with a token on POST /api/v1/signup
The /api/v1/signup endpoint is used to register a user in the system.

The server expects a POST request containing credentials for a user in
JSON format within the request body.

It returns a JSON web token (JWT) in the reponse body per IETF RFC 7519,
with user identification data encoded in the payload.

As implemented here, the server simply reads the username from the
request and inserts that directly into a user object for the reponse; no
user-registry is kept. At present, both the JWT header and signature
fields are dummy strings.