Add SignUpService and a user registration form
SignUpController manages an input form to accept user credentials and
submit them via register() to SignUpService. In turn, the register()
function on the service POSTs these credentials to the server in JSON
format within the request body. The server url and endpoint can be
configured via setters on the provider.

SignUpService expects a JSON web token in the response body per IETF RFC
6750 (and further in RFCs 7235 and 7617). It calls on the AuthToken
service to store the token.

Once registration is complete, the controller broadcasts a success
message to the rest of the application, then redirects the browser to
the dashboard.

In this implementation, the service does not trap error responses from
the server, but always assumes that registration succeeds.